pyproject.toml
[project]
name = "reflex-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "reflex>=0.8.26",
    "supabase>=2.28.0",
    "uvicorn>=0.40.0",
]
